Project 1 completion
Hi there! This is the first commit of this repository. I am following the code of a book called "Arduino Workshop" by John Boxall. In here you will find the schematic for project one and the code. This project creates a blinking wave. This schematic was made using TinkerCAD. Enjoy!